Inline BaseArrayVTable
