+++ encodings/datetime-parts/src/array.rs
@@ -25,7 +25,6 @@ use vortex_array::stats::ArrayStats;
 use vortex_array::stats::StatsSetRef;
 use vortex_array::vtable;
 use vortex_array::vtable::ArrayId;
-use vortex_array::vtable::BaseArrayVTable;
 use vortex_array::vtable::VTable;
 use vortex_array::vtable::ValidityChild;
 use vortex_array::vtable::ValidityVTableFromChild;
@@ -77,14 +76,46 @@ impl VTable for DateTimePartsVTable {
     type Array = DateTimePartsArray;
 
     type Metadata = ProstMetadata<DateTimePartsMetadata>;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = ValidityVTableFromChild;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &DateTimePartsArray) -> usize {
+        array.days.len()
+    }
+
+    fn dtype(array: &DateTimePartsArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &DateTimePartsArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(
+        array: &DateTimePartsArray,
+        state: &mut H,
+        precision: Precision,
+    ) {
+        array.dtype.hash(state);
+        array.days.array_hash(state, precision);
+        array.seconds.array_hash(state, precision);
+        array.subseconds.array_hash(state, precision);
+    }
+
+    fn array_eq(
+        array: &DateTimePartsArray,
+        other: &DateTimePartsArray,
+        precision: Precision,
+    ) -> bool {
+        array.dtype == other.dtype
+            && array.days.array_eq(&other.days, precision)
+            && array.seconds.array_eq(&other.seconds, precision)
+            && array.subseconds.array_eq(&other.subseconds, precision)
     }
 
     fn metadata(array: &DateTimePartsArray) -> VortexResult<Self::Metadata> {
@@ -279,42 +310,6 @@ impl DateTimePartsArray {
 
     pub fn subseconds(&self) -> &ArrayRef {
         &self.subseconds
-    }
-}
-
-impl BaseArrayVTable<DateTimePartsVTable> for DateTimePartsVTable {
-    fn len(array: &DateTimePartsArray) -> usize {
-        array.days.len()
-    }
-
-    fn dtype(array: &DateTimePartsArray) -> &DType {
-        &array.dtype
-    }
-
-    fn stats(array: &DateTimePartsArray) -> StatsSetRef<'_> {
-        array.stats_set.to_ref(array.as_ref())
-    }
-
-    fn array_hash<H: std::hash::Hasher>(
-        array: &DateTimePartsArray,
-        state: &mut H,
-        precision: Precision,
-    ) {
-        array.dtype.hash(state);
-        array.days.array_hash(state, precision);
-        array.seconds.array_hash(state, precision);
-        array.subseconds.array_hash(state, precision);
-    }
-
-    fn array_eq(
-        array: &DateTimePartsArray,
-        other: &DateTimePartsArray,
-        precision: Precision,
-    ) -> bool {
-        array.dtype == other.dtype
-            && array.days.array_eq(&other.days, precision)
-            && array.seconds.array_eq(&other.seconds, precision)
-            && array.subseconds.array_eq(&other.subseconds, precision)
     }
 }
