+++ encodings/fastlanes/src/rle/vtable/mod.rs
@@ -1,16 +1,22 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::hash::Hash;
+
 use prost::Message;
+use vortex_array::ArrayEq;
+use vortex_array::ArrayHash;
 use vortex_array::ArrayRef;
 use vortex_array::ExecutionCtx;
 use vortex_array::IntoArray;
+use vortex_array::Precision;
 use vortex_array::ProstMetadata;
 use vortex_array::buffer::BufferHandle;
 use vortex_array::dtype::DType;
 use vortex_array::dtype::Nullability;
 use vortex_array::dtype::PType;
 use vortex_array::serde::ArrayChildren;
+use vortex_array::stats::StatsSetRef;
 use vortex_array::vtable;
 use vortex_array::vtable::ArrayId;
 use vortex_array::vtable::VTable;
@@ -24,7 +30,6 @@ use crate::rle::array::rle_decompress::rle_decompress;
 use crate::rle::kernel::PARENT_KERNELS;
 use crate::rle::vtable::rules::RULES;
 
-mod array;
 mod operations;
 mod rules;
 mod validity;
@@ -53,13 +58,44 @@ impl VTable for RLEVTable {
 
     type Metadata = ProstMetadata<RLEMetadata>;
 
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = ValidityVTableFromChildSliceHelper;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &RLEArray) -> usize {
+        array.len()
+    }
+
+    fn dtype(array: &RLEArray) -> &DType {
+        array.dtype()
+    }
+
+    fn stats(array: &RLEArray) -> StatsSetRef<'_> {
+        array.stats_set().to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(array: &RLEArray, state: &mut H, precision: Precision) {
+        array.dtype().hash(state);
+        array.values().array_hash(state, precision);
+        array.indices().array_hash(state, precision);
+        array.values_idx_offsets().array_hash(state, precision);
+        array.offset().hash(state);
+        array.len().hash(state);
+    }
+
+    fn array_eq(array: &RLEArray, other: &RLEArray, precision: Precision) -> bool {
+        array.dtype() == other.dtype()
+            && array.values().array_eq(other.values(), precision)
+            && array.indices().array_eq(other.indices(), precision)
+            && array
+                .values_idx_offsets()
+                .array_eq(other.values_idx_offsets(), precision)
+            && array.offset() == other.offset()
+            && array.len() == other.len()
     }
 
     fn reduce_parent(
