+++ vortex-array/src/arrays/chunked/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::hash::Hash;
+
 use itertools::Itertools;
 use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
@@ -13,6 +15,7 @@ use crate::Canonical;
 use crate::EmptyMetadata;
 use crate::ExecutionCtx;
 use crate::IntoArray;
+use crate::Precision;
 use crate::ToCanonical;
 use crate::arrays::ChunkedArray;
 use crate::arrays::PrimitiveArray;
@@ -24,13 +27,14 @@ use crate::builders::ArrayBuilder;
 use crate::dtype::DType;
 use crate::dtype::Nullability;
 use crate::dtype::PType;
+use crate::hash::ArrayEq;
+use crate::hash::ArrayHash;
 use crate::serde::ArrayChildren;
+use crate::stats::StatsSetRef;
 use crate::validity::Validity;
 use crate::vtable;
 use crate::vtable::ArrayId;
 use crate::vtable::VTable;
-
-mod array;
 mod canonical;
 mod operations;
 mod validity;
@@ -49,14 +53,48 @@ impl VTable for ChunkedVTable {
     type Array = ChunkedArray;
 
     type Metadata = EmptyMetadata;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = Self;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &ChunkedArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &ChunkedArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &ChunkedArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(array: &ChunkedArray, state: &mut H, precision: Precision) {
+        array.dtype.hash(state);
+        array.len.hash(state);
+        array.chunk_offsets.as_ref().array_hash(state, precision);
+        for chunk in &array.chunks {
+            chunk.array_hash(state, precision);
+        }
+    }
+
+    fn array_eq(array: &ChunkedArray, other: &ChunkedArray, precision: Precision) -> bool {
+        array.dtype == other.dtype
+            && array.len == other.len
+            && array
+                .chunk_offsets
+                .as_ref()
+                .array_eq(other.chunk_offsets.as_ref(), precision)
+            && array.chunks.len() == other.chunks.len()
+            && array
+                .chunks
+                .iter()
+                .zip(&other.chunks)
+                .all(|(a, b)| a.array_eq(b, precision))
     }
 
     fn metadata(_array: &ChunkedArray) -> VortexResult<Self::Metadata> {
