+++ vortex-array/src/arrays/constant/vtable/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 use std::fmt::Debug;
+use std::hash::Hash;
 
 use vortex_error::VortexResult;
 use vortex_error::vortex_ensure;
@@ -11,6 +12,7 @@ use crate::ArrayRef;
 use crate::EmptyMetadata;
 use crate::ExecutionCtx;
 use crate::IntoArray;
+use crate::Precision;
 use crate::arrays::ConstantArray;
 use crate::arrays::constant::compute::rules::PARENT_RULES;
 use crate::arrays::constant::vtable::canonical::constant_canonicalize;
@@ -19,11 +21,10 @@ use crate::dtype::DType;
 use crate::scalar::Scalar;
 use crate::scalar::ScalarValue;
 use crate::serde::ArrayChildren;
+use crate::stats::StatsSetRef;
 use crate::vtable;
 use crate::vtable::ArrayId;
 use crate::vtable::VTable;
-
-mod array;
 pub(crate) mod canonical;
 mod operations;
 mod validity;
@@ -42,14 +43,37 @@ impl VTable for ConstantVTable {
     type Array = ConstantArray;
 
     type Metadata = EmptyMetadata;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = Self;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &ConstantArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &ConstantArray) -> &DType {
+        array.scalar.dtype()
+    }
+
+    fn stats(array: &ConstantArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(
+        array: &ConstantArray,
+        state: &mut H,
+        _precision: Precision,
+    ) {
+        array.scalar.hash(state);
+        array.len.hash(state);
+    }
+
+    fn array_eq(array: &ConstantArray, other: &ConstantArray, _precision: Precision) -> bool {
+        array.scalar == other.scalar && array.len == other.len
     }
 
     fn metadata(_array: &ConstantArray) -> VortexResult<Self::Metadata> {
