+++ vortex-array/src/arrays/fixed_size_list/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::hash::Hash;
+
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
@@ -10,18 +12,20 @@ use vortex_session::VortexSession;
 use crate::ArrayRef;
 use crate::EmptyMetadata;
 use crate::ExecutionCtx;
+use crate::Precision;
 use crate::arrays::FixedSizeListArray;
 use crate::arrays::fixed_size_list::compute::rules::PARENT_RULES;
 use crate::buffer::BufferHandle;
 use crate::dtype::DType;
+use crate::hash::ArrayEq;
+use crate::hash::ArrayHash;
 use crate::serde::ArrayChildren;
+use crate::stats::StatsSetRef;
 use crate::validity::Validity;
 use crate::vtable;
 use crate::vtable::ArrayId;
 use crate::vtable::VTable;
 use crate::vtable::ValidityVTableFromValidityHelper;
-
-mod array;
 mod kernel;
 mod operations;
 mod validity;
@@ -40,14 +44,48 @@ impl VTable for FixedSizeListVTable {
     type Array = FixedSizeListArray;
 
     type Metadata = EmptyMetadata;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = ValidityVTableFromValidityHelper;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &FixedSizeListArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &FixedSizeListArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &FixedSizeListArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(
+        array: &FixedSizeListArray,
+        state: &mut H,
+        precision: Precision,
+    ) {
+        array.dtype.hash(state);
+        array.elements().array_hash(state, precision);
+        array.list_size().hash(state);
+        array.validity.array_hash(state, precision);
+        array.len.hash(state);
+    }
+
+    fn array_eq(
+        array: &FixedSizeListArray,
+        other: &FixedSizeListArray,
+        precision: Precision,
+    ) -> bool {
+        array.dtype == other.dtype
+            && array.elements().array_eq(other.elements(), precision)
+            && array.list_size() == other.list_size()
+            && array.validity.array_eq(&other.validity, precision)
+            && array.len == other.len
     }
 
     fn reduce_parent(
