+++ vortex-array/src/arrays/listview/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::hash::Hash;
+
 use kernel::PARENT_KERNELS;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
@@ -11,6 +13,7 @@ use vortex_session::VortexSession;
 use crate::ArrayRef;
 use crate::DeserializeMetadata;
 use crate::ExecutionCtx;
+use crate::Precision;
 use crate::ProstMetadata;
 use crate::SerializeMetadata;
 use crate::arrays::ListViewArray;
@@ -19,14 +22,15 @@ use crate::buffer::BufferHandle;
 use crate::dtype::DType;
 use crate::dtype::Nullability;
 use crate::dtype::PType;
+use crate::hash::ArrayEq;
+use crate::hash::ArrayHash;
 use crate::serde::ArrayChildren;
+use crate::stats::StatsSetRef;
 use crate::validity::Validity;
 use crate::vtable;
 use crate::vtable::ArrayId;
 use crate::vtable::VTable;
 use crate::vtable::ValidityVTableFromValidityHelper;
-
-mod array;
 mod kernel;
 mod operations;
 mod validity;
@@ -55,14 +59,45 @@ impl VTable for ListViewVTable {
     type Array = ListViewArray;
 
     type Metadata = ProstMetadata<ListViewMetadata>;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = ValidityVTableFromValidityHelper;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &ListViewArray) -> usize {
+        debug_assert_eq!(array.offsets().len(), array.sizes().len());
+        array.offsets().len()
+    }
+
+    fn dtype(array: &ListViewArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &ListViewArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(
+        array: &ListViewArray,
+        state: &mut H,
+        precision: Precision,
+    ) {
+        array.dtype.hash(state);
+        array.elements().array_hash(state, precision);
+        array.offsets().array_hash(state, precision);
+        array.sizes().array_hash(state, precision);
+        array.validity.array_hash(state, precision);
+    }
+
+    fn array_eq(array: &ListViewArray, other: &ListViewArray, precision: Precision) -> bool {
+        array.dtype == other.dtype
+            && array.elements().array_eq(other.elements(), precision)
+            && array.offsets().array_eq(other.offsets(), precision)
+            && array.sizes().array_eq(other.sizes(), precision)
+            && array.validity.array_eq(&other.validity, precision)
     }
 
     fn metadata(array: &ListViewArray) -> VortexResult<Self::Metadata> {
