+++ vortex-array/src/arrays/varbinview/vtable/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::hash::Hash;
 use std::mem::size_of;
 use std::sync::Arc;
 
@@ -14,19 +15,21 @@ use vortex_session::VortexSession;
 use crate::ArrayRef;
 use crate::EmptyMetadata;
 use crate::ExecutionCtx;
+use crate::Precision;
 use crate::arrays::BinaryView;
 use crate::arrays::varbinview::VarBinViewArray;
 use crate::arrays::varbinview::compute::rules::PARENT_RULES;
 use crate::buffer::BufferHandle;
 use crate::dtype::DType;
+use crate::hash::ArrayEq;
+use crate::hash::ArrayHash;
 use crate::serde::ArrayChildren;
+use crate::stats::StatsSetRef;
 use crate::validity::Validity;
 use crate::vtable;
 use crate::vtable::ArrayId;
 use crate::vtable::VTable;
 use crate::vtable::ValidityVTableFromValidityHelper;
-
-mod array;
 mod kernel;
 mod operations;
 mod validity;
@@ -45,14 +48,49 @@ impl VTable for VarBinViewVTable {
     type Array = VarBinViewArray;
 
     type Metadata = EmptyMetadata;
-
-    type ArrayVTable = Self;
     type OperationsVTable = Self;
     type ValidityVTable = ValidityVTableFromValidityHelper;
     type VisitorVTable = Self;
 
     fn id(_array: &Self::Array) -> ArrayId {
         Self::ID
+    }
+
+    fn len(array: &VarBinViewArray) -> usize {
+        array.views_handle().len() / size_of::<BinaryView>()
+    }
+
+    fn dtype(array: &VarBinViewArray) -> &DType {
+        &array.dtype
+    }
+
+    fn stats(array: &VarBinViewArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+
+    fn array_hash<H: std::hash::Hasher>(
+        array: &VarBinViewArray,
+        state: &mut H,
+        precision: Precision,
+    ) {
+        array.dtype.hash(state);
+        for buffer in array.buffers.iter() {
+            buffer.array_hash(state, precision);
+        }
+        array.views.array_hash(state, precision);
+        array.validity.array_hash(state, precision);
+    }
+
+    fn array_eq(array: &VarBinViewArray, other: &VarBinViewArray, precision: Precision) -> bool {
+        array.dtype == other.dtype
+            && array.buffers.len() == other.buffers.len()
+            && array
+                .buffers
+                .iter()
+                .zip(other.buffers.iter())
+                .all(|(a, b)| a.array_eq(b, precision))
+            && array.views.array_eq(&other.views, precision)
+            && array.validity.array_eq(&other.validity, precision)
     }
 
     fn metadata(_array: &VarBinViewArray) -> VortexResult<Self::Metadata> {
